ADQL type calculus moves from adqlglue to the adql package; this means, in particular, that FieldInfos now have a type attribute and must be constructed with a type.
getSubsumingType and helpers moves from typesystems to a module of its
own and learns arrays and geometric types.

+++ tests/taptest.py
@@ -265,7 +265,7 @@ class SimpleRunnerTest(testhelpers.VerboseTest):
 		self.assertEqual(fields[0].datatype, "double")
 		self.assertEqual(fields[1].datatype, "char")
 		self.assertEqual(fields[1].xtype, "adql:POINT")
-		self.assertEqual(fields[2].datatype, "float")
+		self.assertEqual(fields[2].datatype, "double")
 
 
 class TAPSchemaTest(testhelpers.VerboseTest):
